// src/enums/tabTypes.ts
export enum MenuTabs {
    DOCTOR = "Doctor",
    WALLET = "Wallet",
    REFUND = "Refund",
    COUPONS = "Coupons",
    CONCERNS = "Concerns",
    REFERRAL = "Referral",
    PATIENTS = "Patients",
    DASHBOARD = "Dashboard",
    APPOINTMENT = "Appointment",
    SPECIALTIES = "Specialties",
    INGREDIENTS = "Ingredients",
    CUSTOMIZATIONS = "Customizations",
    ADD_INGREDIENTS = "Add Ingredients",
    INGREDIENTS_LIST = "Ingredients List"
}

export enum TooltipPlacement {
  RIGHT = "right",
  TOP = "top"
}

export enum StatusTypes {
  ACTIVE = "active",
  INACTIVE = "inactive"
}